Feature/api scheduler 15 gunwoong (#28)
* feat: api

* feat: api 스케줄링

* feat: spot service inteface

* test: remove legacy test

* feat: apply open meteo

* test: apply api test

+++ src/main/java/sevenstar/marineleisure/global/api/khoa/dto/item/FishingItem.java
@@ -13,7 +13,7 @@ public class FishingItem implements KhoaItem {
 	private String predcYmd;
 	private String predcNoonSeCd;
 	private String seafsTgfshNm;
-	private float tdlvHrScr;
+	private int tdlvHrScr;
 	private float minWvhgt;
 	private float maxWvhgt;
 	private float minWtem;
